requirements.txt
absl-py==1.3.0
aio-pika==8.2.3
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiormq==6.4.2
aiosignal==1.3.1
altair==4.2.2
anyascii==0.3.1
anyio==3.6.2
aplus==0.11.0
APScheduler==3.9.1.post1
argcomplete==1.10.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astropy==5.2.2
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Automat==22.10.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.8.2
bidict==0.22.1
blake3==0.3.3
bleach==5.0.1
blinker==1.6.2
blis==0.7.9
boto3==1.26.126
botocore==1.29.126
bqplot==0.12.39
branca==0.6.0
CacheControl==0.12.11
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chart-studio==1.1.0
ci-info==0.3.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.5
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
colorlover==0.3.0
Columnar==1.4.1
compressed-rtf==1.0.6
comtypes==1.1.14
confection==0.0.3
configobj==5.0.8
configparser==5.3.0
confluent-kafka==1.9.2
constantly==15.1.0
construct==2.10.68
contourpy==1.0.5
contractions==0.1.73
cryptography==40.0.1
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.7
Cython==0.29.28
dask==2022.10.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
dnspython==1.16.0
docopt==0.6.2
docx2txt==0.8
ebcdic==1.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
etelemetry==0.3.0
executing==1.1.1
extract-msg==0.28.7
fastapi==0.95.2
fastjsonschema==2.16.2
fbmessenger==6.0.0
filelock==3.9.0
findspark==2.0.1
fire==0.5.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.3.3
fsspec==2023.4.0
funcy==2.0
future==0.18.3
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
httplib2==0.22.0
httptools==0.5.0
huggingface-hub==0.12.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imageio==2.29.0
IMAPClient==2.1.0
imbalanced-learn==0.10.1
importlib-metadata==6.6.0
imutils==0.5.4
incremental==22.10.0
install==1.3.5
ipydatawidgets==4.3.3
ipykernel==6.16.1
ipyleaflet==0.17.2
ipympl==0.9.3
ipython==8.5.0
ipython-genutils==0.2.0
ipyvolume==0.6.1
ipyvue==1.9.0
ipyvuetify==1.8.10
ipywebrtc==0.6.0
ipywidgets==8.0.3
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
keras==2.11.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
lazy_loader==0.2
libclang==14.0.6
lime==0.2.0.1
llvmlite==0.40.0
locket==1.0.0
looseversion==1.1.2
lxml==4.9.1
magic-filter==1.0.9
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mattermostwrapper==2.2
mdurl==0.1.2
mistune==2.0.4
mpmath==1.3.0
msgpack==1.0.5
mtcnn==0.1.1
multidict==5.2.0
murmurhash==1.0.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
nibabel==5.1.0
nipype==1.8.6
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
num2words==0.5.12
numba==0.57.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
openai==0.27.7
opencv-python==4.7.0.68
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathlib==1.0.1
pathy==0.10.1
patsy==0.5.3
pdfminer.six==20191110
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.2
playsound==1.2.2
plotly==5.11.0
pluggy==1.0.0
ply==3.11
portalocker==2.7.0
preshed==3.0.8
progressbar2==4.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.28
protobuf==3.20.1
prov==2.0.0
psutil==5.9.3
psycopg2-binary==2.9.6
pure-eval==0.2.2
Py-Dictionary==4.1.4
pyahocorasick==1.4.4
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.16.0
pydantic==1.10.2
pydeck==0.8.1b0
pydot==1.4.2
pydotplus==2.0.2
pyerfa==2.0.0.3
Pygments==2.13.0
PyJWT==2.6.0
pykwalify==1.8.0
pyLDAvis==3.4.1
pymongo==3.10.1
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==2.12.1
pypiwin32==223
pyreadline3==3.4.1
pyrsistent==0.19.3
pysmi==0.3.4
pysnmp==4.4.12
PySocks==1.7.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-json-logger==2.0.4
python-pptx==0.6.21
python-socketio==5.8.0
python-utils==3.5.2
pythreejs==2.4.2
pyttsx3==2.90
pytube==15.0.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==304
pywinpty==2.0.10
pyxnat==1.5
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
questionary==1.10.0
randomname==0.1.5
rasa-sdk==3.5.1
rdflib==6.3.2
redis==4.5.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rocketchat-API==1.28.1
rsa==4.9
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-image==0.20.0
scikit-learn==1.1.2
scipy==1.8.1
seaborn==0.12.2
Send2Trash==1.8.0
sentencepiece==0.1.99
sentry-sdk==1.14.0
simplejson==3.19.1
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.21.3
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.3
SpeechRecognition==3.8.1
SQLAlchemy==1.4.48
srsly==2.4.5
stack-data==0.5.1
starlette==0.27.0
statsmodels==0.13.5
streamlit==1.22.0
string-color==1.2.3
sympy==1.12
tabula-py==2.7.0
tabulate==0.9.0
tarsafe==0.0.3
tenacity==8.1.0
tensor==0.3.6
tensorboard==2.11.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.17.1
terminaltables==3.1.10
textblob==0.17.1
textract-trp==0.1.3
textsearch==0.0.24
tf-estimator-nightly==2.14.0.dev2023060108
tflearn==0.5.0
thinc==8.1.5
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchdata==0.6.1
torchtext==0.15.2
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
traits==6.3.2
traittypes==0.2.1
transformers==4.26.0
twilio==7.14.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typeguard==2.13.3
typer==0.7.0
typing-utils==0.1.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
uri-template==1.2.0
urllib3==1.26.12
uvicorn==0.21.1
vaex==4.16.0
vaex-astro==0.9.3
vaex-core==4.16.1
vaex-hdf5==0.14.1
vaex-jupyter==0.8.1
vaex-ml==0.18.1
vaex-server==0.8.1
vaex-viz==0.5.4
validators==0.20.0
wasabi==0.10.1
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
webexteamssdk==1.6.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.4
wordcloud==1.8.2.2
wrapt==1.14.1
xarray==2023.4.2
xgboost==1.7.5
xlrd==1.2.0
XlsxWriter==3.0.9
xyzservices==2023.2.0
yarl==1.9.2
zipp==3.15.0
zope.interface==5.5.2
